Move legacy thread proxying system into seperate source and header. NFC
This is the first step towards removing the internal use of the old
proxying system.  The primary user of which is the JS library proxying
system.

The legacy proxying system is already layered on top of the new proxying
system so removing it should be pure win.
